Changed Select command to have both SelectNextQuestion or SelectPreviousQuestion with different CanSelect method.
For the SelectPreviousQuestion command, there is one specific method for determining whether you can select the previous question. For the SelectNextQuestion command, there is a general method that determines whether you can select the next question, or modifiy the questions(randomize,redo).

Implemented Reveal for the selection gamemode, and renamed some of the methods. Also added a new method TryResetStatus